+++ packages/sdk/src/evm/core/classes/index.ts
@@ -8,11 +8,7 @@ export * from "./drop-claim-conditions";
 export * from "./drop-erc1155-claim-conditions";
 export * from "./drop-erc1155-history";
 export * from "./erc-20";
-export * from "./erc-20-mintable";
-export * from "./erc-20-batch-mintable";
 export * from "./erc-20-history";
-export * from "./erc-20-signature-mintable";
-export * from "./erc-20-burnable";
 export * from "./erc-20-standard";
 export * from "./erc-721";
 export * from "./erc-721-lazymintable";
